Move utils to pkg directory
