+++ digitalstrom/circuits.go
@@ -2,7 +2,7 @@ package digitalstrom
 
 import (
 	"github.com/gaetancollaud/digitalstrom-mqtt/pkg/digitalstrom"
-	"github.com/gaetancollaud/digitalstrom-mqtt/utils"
+	"github.com/gaetancollaud/digitalstrom-mqtt/pkg/utils"
 	"github.com/rs/zerolog/log"
 )
