+++ digitalstrom/devices.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	"github.com/gaetancollaud/digitalstrom-mqtt/pkg/digitalstrom"
-	"github.com/gaetancollaud/digitalstrom-mqtt/utils"
+	"github.com/gaetancollaud/digitalstrom-mqtt/pkg/utils"
 	"github.com/rs/zerolog/log"
 )
